Move new Trust methods into TrustService and use AcademyRepository
Also refactor some tests that were modified by the move
