+++ DfE.FindInformationAcademiesTrusts.Data.AcademiesDb/TrustProvider.cs
@@ -1,7 +1,5 @@
 using DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Contexts;
-using DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Extensions;
 using DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Factories;
-using DfE.FindInformationAcademiesTrusts.Data.Dto;
 using Microsoft.EntityFrameworkCore;
 
 namespace DfE.FindInformationAcademiesTrusts.Data.AcademiesDb;
@@ -14,67 +12,6 @@ public class TrustProvider(
     IPersonFactory personFactory)
     : ITrustProvider
 {
-    public async Task<TrustSummaryDto?> GetTrustSummaryAsync(string uid)
-    {
-        var details = await academiesDbContext.Groups
-            .Where(g => g.GroupUid == uid)
-            .Select(g => new { Name = g.GroupName ?? string.Empty, Type = g.GroupType ?? string.Empty })
-            .SingleOrDefaultAsync();
-
-        if (details is null) return null;
-
-        var count = await academiesDbContext.GiasGroupLinks.CountAsync(gl => gl.GroupUid == uid && gl.Urn != null);
-
-        return new TrustSummaryDto(uid, details.Name, details.Type, count);
-    }
-
-    public async Task<TrustDetailsDto> GetTrustDetailsAsync(string uid)
-    {
-        var regionAndTerritory = await academiesDbContext.MstrTrusts
-            .Where(m => m.GroupUid == uid)
-            .Select(m => m.GORregion)
-            .SingleOrDefaultAsync() ?? string.Empty;
-
-        var singleAcademyUrn = await GetUrnForSingleAcademyTrust(uid);
-
-        var giasGroup = await academiesDbContext.Groups
-            .Where(g => g.GroupUid == uid)
-            .Select(giasGroup => new
-            {
-                giasGroup.GroupUid,
-                giasGroup.GroupId,
-                giasGroup.Ukprn,
-                giasGroup.CompaniesHouseNumber,
-                giasGroup.GroupType,
-                giasGroup.GroupContactStreet,
-                giasGroup.GroupContactLocality,
-                giasGroup.GroupContactTown,
-                giasGroup.GroupContactPostcode,
-                giasGroup.IncorporatedOnOpenDate
-            })
-            .SingleAsync();
-
-        var trustDetailsDto = new TrustDetailsDto(
-            giasGroup.GroupUid!, //Searched by this field so it must be present
-            giasGroup.GroupId,
-            giasGroup.Ukprn,
-            giasGroup.CompaniesHouseNumber,
-            giasGroup.GroupType!, //Enforced by EF filter
-            string.Join(", ", new[]
-            {
-                giasGroup.GroupContactStreet,
-                giasGroup.GroupContactLocality,
-                giasGroup.GroupContactTown,
-                giasGroup.GroupContactPostcode
-            }.Where(s => !string.IsNullOrWhiteSpace(s))),
-            regionAndTerritory,
-            singleAcademyUrn,
-            giasGroup.IncorporatedOnOpenDate.ParseAsNullableDate()
-        );
-
-        return trustDetailsDto;
-    }
-
     public async Task<Trust?> GetTrustByUidAsync(string uid)
     {
         var giasGroup = await academiesDbContext.Groups.SingleOrDefaultAsync(g => g.GroupUid == uid);
@@ -117,15 +54,6 @@ public class TrustProvider(
                 mtg => mtg.Gid,
                 (gg, mtg) => governorFactory.CreateFrom(gg, mtg))
             .ToArrayAsync();
-    }
-
-    private async Task<string?> GetUrnForSingleAcademyTrust(string uid)
-    {
-        return await academiesDbContext.GiasGroupLinks
-            .Where(gl => gl.GroupUid == uid
-                         && gl.GroupType == "Single-academy trust")
-            .Select(gl => gl.Urn)
-            .FirstOrDefaultAsync();
     }
 
     private async Task<Academy[]> GetAcademiesLinkedTo(string uid)
