+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/Pages/Trusts/Academies/AcademiesDetailsModelTests.cs
@@ -1,4 +1,5 @@
 using DfE.FindInformationAcademiesTrusts.Data;
+using DfE.FindInformationAcademiesTrusts.Data.AcademiesDb;
 using DfE.FindInformationAcademiesTrusts.Data.Dto;
 using DfE.FindInformationAcademiesTrusts.Data.UnitTests.Mocks;
 using DfE.FindInformationAcademiesTrusts.Pages;
@@ -12,21 +13,22 @@ public class AcademiesDetailsModelTests
 {
     private readonly AcademiesDetailsModel _sut;
     private readonly Mock<IOtherServicesLinkBuilder> _mockLinkBuilder = new();
-    private readonly Mock<ITrustProvider> _mockTrustProvider;
-    private readonly MockDataSourceProvider _mockDataSourceProvider;
+    private readonly Mock<ITrustProvider> _mockTrustProvider = new();
+    private readonly Mock<ITrustService> _mockTrustRepository = new();
+    private readonly MockDataSourceProvider _mockDataSourceProvider = new();
 
     public AcademiesDetailsModelTests()
     {
         MockLogger<AcademiesDetailsModel> logger = new();
         var dummyTrust = DummyTrustFactory.GetDummyTrust("1234");
-        _mockTrustProvider = new Mock<ITrustProvider>();
-        _mockDataSourceProvider = new MockDataSourceProvider();
+
         _mockTrustProvider.Setup(tp => tp.GetTrustByUidAsync("1234")).ReturnsAsync(dummyTrust);
-        _mockTrustProvider.Setup(tp => tp.GetTrustSummaryAsync(dummyTrust.Uid))
+        _mockTrustRepository.Setup(t => t.GetTrustSummaryAsync(dummyTrust.Uid))
             .ReturnsAsync(new TrustSummaryDto(dummyTrust.Uid, dummyTrust.Name, dummyTrust.Type,
                 dummyTrust.Academies.Length));
+
         _sut = new AcademiesDetailsModel(_mockTrustProvider.Object, _mockDataSourceProvider.Object,
-            _mockLinkBuilder.Object, logger.Object) { Uid = "1234" };
+            _mockLinkBuilder.Object, logger.Object, _mockTrustRepository.Object) { Uid = "1234" };
     }
 
     [Fact]
@@ -56,7 +58,7 @@ public class AcademiesDetailsModelTests
     [Fact]
     public async Task OnGetAsync_returns_NotFoundResult_if_Trust_is_not_found()
     {
-        _mockTrustProvider.Setup(tp => tp.GetTrustSummaryAsync("1234")).ReturnsAsync((TrustSummaryDto?)null);
+        _mockTrustRepository.Setup(t => t.GetTrustSummaryAsync("1234")).ReturnsAsync((TrustSummaryDto?)null);
         var result = await _sut.OnGetAsync();
         result.Should().BeOfType<NotFoundResult>();
     }
